refactor: Exception Handler 위치 변경
