+++ src/main/java/techcourse/myblog/web/controller/UserControllerExceptionHandler.java
@@ -1,9 +1,11 @@
 package techcourse.myblog.web.controller;
 
+import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
+import techcourse.myblog.service.dto.UserDto;
 import techcourse.myblog.support.exception.*;
 
 @ControllerAdvice
@@ -36,5 +38,12 @@ public class UserControllerExceptionHandler {
     public RedirectView handleCommentFail(InvalidCommentException e, RedirectAttributes redirectAttributes) {
         redirectAttributes.addFlashAttribute("error", e.getMessage());
         return new RedirectView("/");
+    }
+
+    @ExceptionHandler(BindException.class)
+    public RedirectView handleBindError(BindException e, RedirectAttributes redirectAttributes) {
+        redirectAttributes.addFlashAttribute("userDto", new UserDto("", "", ""));
+        redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.userDto", e.getBindingResult());
+        return new RedirectView(e.getObjectName());
     }
 }
